+++ examples/07_load_finetune.yaml
@@ -60,12 +60,12 @@ exp2-finetune-model: !LoadSerialized
   filename: examples/models/exp1-pretrain-model.mod
   path: ''
   overwrite: # list of [path, value] pairs. Value can be scalar or an arbitrary object
-  - path: train.trainer
+  - path: training.trainer
     val: !AdamTrainer
         alpha: 0.0002
   - path: exp_global.dropout
     val: 0.5
-  - path: train.dev_zero
+  - path: training.dev_zero
     val: True
   - path: status
     val: null
