+++ examples/09_programmatic.py
@@ -34,7 +34,7 @@ from xnmt.optimizers import AdamTrainer
 from xnmt.param_collections import ParamManager
 from xnmt.persistence import save_to_file
 import xnmt.tee
-from xnmt.train.regimens import SimpleTrainingRegimen
+from xnmt.training.regimens import SimpleTrainingRegimen
 from xnmt.models.translators import DefaultTranslator
 from xnmt.vocabs import Vocab
