+++ examples/15_score.yaml
@@ -56,7 +56,7 @@ exp1-model: !Experiment
 exp2-score: !LoadSerialized
   filename: examples/output/exp1-model.mod
   overwrite:
-  - path: train
+  - path: training
     val: ~
   - path: model.inference
     val: !AutoRegressiveInference
