+++ xnmt/input_readers/__init__.py
@@ -1,0 +1,10 @@
+from xnmt.input_readers.text.char_from_word import CharFromWordTextReader
+from xnmt.input_readers.text.id import IDReader
+from xnmt.input_readers.text.plain import PlainTextReader
+from xnmt.input_readers.text.raml import RamlTextReader
+from xnmt.input_readers.text.sentpiece import SentencePieceTextReader
+from xnmt.input_readers.compound import CompoundReader
+from xnmt.input_readers.h5 import H5Reader
+from xnmt.input_readers.npz import NpzReader
+from xnmt.input_readers.input_reader import InputReader
+from xnmt.input_readers.utils import read_parallel_corpus
