+++ xnmt/input_readers/input_reader.py
@@ -1,0 +1,36 @@
+import numbers
+
+from xnmt import sent
+from typing import Sequence, Iterator
+
+
+class InputReader(object):
+  """
+  A base class to read in a file and turn it into an input
+  """
+  def read_sents(self, filename: str, filter_ids: Sequence[numbers.Integral] = None) -> Iterator[sent.Sentence]:
+    """
+    Read sentences and return an iterator.
+
+    Args:
+      filename: data file
+      filter_ids: only read sentences with these ids (0-indexed)
+    Returns: iterator over sentences from filename
+    """
+    return self.iterate_filtered(filename, filter_ids)
+
+  def count_sents(self, filename: str) -> int:
+    """
+    Count the number of sentences in a data file.
+
+    Args:
+      filename: data file
+    Returns: number of sentences in the data file
+    """
+    raise RuntimeError("Input readers must implement the count_sents function")
+
+  def needs_reload(self) -> bool:
+    """
+    Overwrite this method if data needs to be reload for each epoch
+    """
+    return False
