+++ xnmt/models/translators/__init__.py
@@ -5,3 +5,6 @@ TranslatorOutput = namedtuple('TranslatorOutput', ['state', 'logsoftmax', 'atten
 from xnmt.models.translators.auto_regressive import AutoRegressiveTranslator
 from xnmt.models.translators.default_translator import DefaultTranslator
 from xnmt.models.translators.ensemble_translator import EnsembleTranslator
+
+
+
