+++ xnmt/rl/policy_gradient.py
@@ -8,6 +8,7 @@ from xnmt.persistence import Ref, bare, Serializable, serializable_init
 from xnmt.losses import FactoredLossExpr
 from xnmt.param_initializers import GlorotInitializer, ZeroInitializer
 
+
 class PolicyGradient(Serializable):
   """
   (Sequence) policy gradient class. It holds a policy network that will perform a linear regression to the output_dim decision labels.
@@ -25,39 +26,56 @@ class PolicyGradient(Serializable):
     use_baseline: Whether to turn on baseline reward discounting or not.
     weight: The weight of the reinforce_loss.
     output_dim: The size of the predicitions.
-  """ 
-
-
+  """
   yaml_tag = '!PolicyGradient'
+
   @serializable_init
   @register_xnmt_handler
-  def __init__(self, policy_network=None,
-                     baseline=None,
-                     z_normalization=True,
-                     conf_penalty=None,
-                     weight=1.0,
-                     input_dim=Ref("exp_global.default_layer_dim"),
-                     output_dim=2,
-                     param_init=Ref("exp_global.param_init", default=bare(GlorotInitializer)),
-                     bias_init=Ref("exp_global.bias_init", default=bare(ZeroInitializer))):
+  def __init__(self,
+               policy_network=None,
+               baseline=None,
+               conf_penalty=None,
+               weight=1.0,
+               input_dim=Ref("exp_global.default_layer_dim"),
+               output_dim=2,
+               z_normalization=True,
+               param_init=Ref("exp_global.param_init", default=bare(GlorotInitializer)),
+               bias_init=Ref("exp_global.bias_init", default=bare(ZeroInitializer))):
     self.input_dim = input_dim
     self.policy_network = self.add_serializable_component("policy_network",
                                                            policy_network,
-                                                           lambda: Linear(input_dim=self.input_dim, output_dim=output_dim,
-                                                                          param_init=param_init, bias_init=bias_init))
+                                                           lambda: Linear(input_dim=self.input_dim,
+                                                                          output_dim=output_dim,
+                                                                          param_init=param_init,
+                                                                          bias_init=bias_init))
     self.baseline = self.add_serializable_component("baseline", baseline,
-                                                    lambda: Linear(input_dim=self.input_dim, output_dim=1,
-                                                                   param_init=param_init, bias_init=bias_init))
+                                                    lambda: Linear(input_dim=self.input_dim,
+                                                                   output_dim=1,
+                                                                   param_init=param_init,
+                                                                   bias_init=bias_init))
 
-    self.confidence_penalty = self.add_serializable_component("conf_penalty", conf_penalty, lambda: conf_penalty) if conf_penalty is not None else None
+    self.confidence_penalty = self.add_serializable_component("conf_penalty",
+                                                              conf_penalty,
+                                                              lambda: conf_penalty)
     self.weight = weight
+    # State of the object
+    self.sampling_action = None
+    self.policy_lls = []
+    self.actions = []
+    self.states = []
+    self.baseline_input = None
+    self.valid_pos = None
+    # Deprecated
+    self.z_normalization = z_normalization
 
   """
   state: Input state.
   argmax: Whether to perform argmax or sampling.
-  sample_pp: Stands for sample post_processing. Every time the sample are being drawn, this method will be invoked with sample_pp(sample).
-  predefined_actions: Whether to forcefully the network to assign the action value to some predefined actions. This predefined actions can be 
-                      from the gold distribution or some probability priors. It should be calculated from the outside.
+  sample_pp: Stands for sample post_processing. Every time the sample are being drawn, this method will be 
+             invoked with sample_pp(sample).
+  predefined_actions: Whether to forcefully the network to assign the action value to some predefined actions. 
+                      This predefined actions can be from the gold distribution or some probability priors. 
+                      It should be calculated from the outside.
   """
   def sample_action(self, state, argmax=False, sample_pp=None, predefined_actions=None):
     policy = dy.log_softmax(self.policy_network.transform(state))
@@ -83,21 +101,20 @@ class PolicyGradient(Serializable):
   """
   def calc_loss(self, policy_reward):
     loss = FactoredLossExpr()
-    ## Calculate baseline
+    # Calculate baseline
     pred_reward, baseline_loss = self.calc_baseline_loss(policy_reward)
     rewards = [policy_reward - pw_i for pw_i in pred_reward]
     loss.add_loss("rl_baseline", baseline_loss)
-    ## Z-Normalization
+    # Z-Normalization
     rewards = dy.concatenate(rewards, d=0)
-    dim, batch_size = rewards.dim()
     rewards_mean = dy.mean_dim(rewards, [0], False)
     rewards_std = dy.std_dim(rewards, [0], False)
     rewards = dy.cdiv(rewards - rewards_mean, rewards_std+1e-10)
-    ## Calculate Confidence Penalty
+    # Calculate Confidence Penalty
     if self.confidence_penalty:
       cp_loss = self.confidence_penalty.calc_loss(self.policy_lls)
       loss.add_loss("rl_confpen", cp_loss)
-    ## Calculate Reinforce Loss
+    # Calculate Reinforce Loss
     reinf_loss = []
     # Loop through all action in one sequence
     for i, (policy, action) in enumerate(zip(self.policy_lls, self.actions)):
@@ -109,7 +126,7 @@ class PolicyGradient(Serializable):
         reward = dy.pick_batch_elems(reward, self.valid_pos[i])
       reinf_loss.append(dy.sum_batches(-ll * reward))
     loss.add_loss("rl_reinf", self.weight * dy.esum(reinf_loss))
-    ## the composed losses
+    # the composed losses
     return loss
 
   def shared_params(self):
@@ -118,9 +135,9 @@ class PolicyGradient(Serializable):
 
   @handle_xnmt_event
   def on_start_sent(self, src_sent):
-    self.policy_lls = []
-    self.actions = []
-    self.states = []
+    self.policy_lls.clear()
+    self.actions.clear()
+    self.states.clear()
     self.baseline_input = None
     self.valid_pos = src_sent.mask.get_valid_position() if src_sent.mask else None 
     self.sampling_action = self.SamplingAction.NONE
